Refactor SVPWM scaling and modulation index handling; add line-to-line voltage simulation script

+++ fw/Induction-Machine-FOC/Core/Inc/sine_op.h
@@ -13,6 +13,6 @@
 
 void build_sine_lut(void);
 int16_t get_sine_q15(uint32_t phase);
-uint32_t sine_to_cmp(int16_t s_q15, uint32_t per, int16_t m_q15);
+uint32_t sine_to_cmp(int16_t s_q15, uint32_t per, int32_t m_q15);
 
 #endif // SINE_OP_H
